// backend/src/common/enums/role.enum.ts
// src/common/enums/role.enum.ts
export enum Role {
  CUSTOMER = 'CUSTOMER', // Klient i zakonshëm - mund të rezervojë
  STAFF = 'STAFF', // Staf - mund të menaxhojë rezervimet
  ADMIN = 'ADMIN', // Admin - mund të menaxhojë agjencinë
  SUPER_ADMIN = 'SUPER_ADMIN', // Super Admin - mund të menaxhojë të gjitha agjencitë
}